+++ src/components/POLine/Location/LocationForm.js
@@ -4,24 +4,34 @@ import PropTypes from 'prop-types';
 import { ifDisabledToChangePaymentInfo, isWorkflowStatusIsPending } from '../../PurchaseOrder/util';
 import { FieldsLocation } from '../../../common/POLFields';
 
-const LocationForm = ({ order, locations, isPackage }) => {
+const LocationForm = ({
+  changeLocation,
+  isPackage,
+  locationIds,
+  locations,
+  order,
+}) => {
   const isPostPendingOrder = !isWorkflowStatusIsPending(order);
   const isDisabledToChangePaymentInfo = ifDisabledToChangePaymentInfo(order);
 
   return (
     <FieldsLocation
-      locations={locations}
+      changeLocation={changeLocation}
       disabled={isPostPendingOrder}
       isDisabledToChangePaymentInfo={isDisabledToChangePaymentInfo}
+      locationIds={locationIds}
+      locations={locations}
       withValidation={!isPackage}
     />
   );
 };
 
 LocationForm.propTypes = {
+  changeLocation: PropTypes.func.isRequired,
+  isPackage: PropTypes.bool,
+  locationIds: PropTypes.arrayOf(PropTypes.string),
   locations: PropTypes.arrayOf(PropTypes.object),
   order: PropTypes.object.isRequired,
-  isPackage: PropTypes.bool,
 };
 
 LocationForm.defaultProps = {