+++ src/components/POLine/POLineForm.js
@@ -26,7 +26,6 @@ import { ViewMetaData } from '@folio/stripes/smart-components';
 import stripesForm from '@folio/stripes/form';
 import {
   FundDistributionFields,
-  getLocationOptions,
 } from '@folio/stripes-acq-components';
 
 import {
@@ -207,6 +206,12 @@ class POLineForm extends Component {
     this.setState({ sections });
   };
 
+  changeLocation = (location, fieldName) => {
+    const { change, dispatch } = this.props;
+
+    dispatch(change(fieldName, location.id));
+  };
+
   render() {
     const {
       change,
@@ -240,7 +245,8 @@ class POLineForm extends Component {
     const identifierTypes = getIdentifierTypesForSelect(parentResources);
     const contributorNameTypes = getContributorNameTypesForSelect(parentResources);
     const orderTemplates = getOrderTemplatesForSelect(parentResources);
-    const locations = getLocationOptions(parentResources?.locations?.records || []);
+    const locations = parentResources?.locations?.records;
+    const locationIds = locations?.map(({ id }) => id);
     const isDisabledToChangePaymentInfo = ifDisabledToChangePaymentInfo(order);
     const estimatedPrice = calculateEstimatedPrice(formValues, stripes.currency);
     const { accounts } = vendor;
@@ -361,9 +367,12 @@ class POLineForm extends Component {
                       id={ACCORDION_ID.location}
                     >
                       <LocationForm
+                        changeLocation={this.changeLocation}
+                        formValues={formValues}
+                        isPackage={isPackage}
+                        locationIds={locationIds}
                         locations={locations}
                         order={order}
-                        isPackage={isPackage}
                       />
                     </Accordion>
                     {showPhresources && (